part3/requirements.txt
flask
flask-restx
flask-bcrypt
flask-jwt-extended
lask_sqlalchemy